Move forgot password under user specs
