+++ app/controllers/users_controller.rb
@@ -18,6 +18,15 @@ class UsersController < ApplicationController
     render json: user
   end
 
+  def forgot_password
+    user = User.find_by(email: params[:user][:email])
+    if user && user.forgot_password!
+      render json: user
+    else
+      render_no_such_email_error
+    end
+  end
+
   def reset_password
     user = find_user_by_confirmation_token
     if user && user.update_password(params[:user][:password])
@@ -31,6 +40,10 @@ class UsersController < ApplicationController
 
   def permitted_params
     params.require(:user).permit(:email, :username, :password, :confirmation_token,)
+  end
+
+  def render_no_such_email_error
+    render json: {errors: {email: ["doesn't exist"]}}, status: 422
   end
 
   def render_could_not_reset_password_error
